fix: format string （.*s) expect int type for describe length

+++ libx/dump.c
@@ -385,7 +385,7 @@ void x_dump_free(x_dump *dmp)
 static int write_file_cb(const x_uchar *buf, size_t len, void *ctx)
 {
 	FILE *fp = ctx;
-	return x_fprintf(fp, x_u("%.*s"), len, buf);
+	return x_fprintf(fp, x_u("%.*s"), (int)len, buf);
 }
 
 static int indent_check_cb(const x_uchar *buf, size_t len, void *ctx)
